etl/requirements.txt
requests
pandas
numpy
psycopg2-binary